% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cervical_cancer}
\alias{Cervical_cancer}
\title{Cervical_cancer-This data set predicts a patient's risk of cervical cancer based on behavior reports}
\format{
An object of class \code{data.frame} with 858 rows and 34 columns.
}
\source{
https://archive.ics.uci.edu/dataset/383/cervical+cancer+risk+factors
}
\usage{
Cervical_cancer
}
\description{
"The dataset was collected at 'Hospital Universitario de Caracas' in Caracas, Venezuela. The dataset comprises demographic information, habits, and historic medical records of 858 patients.
Several patients decided not to answer some of the questions because of privacy concerns (missing values)."
I cleaned up the data so there are no missing data points, nor any NAs.
}
\keyword{datasets}
