% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Diabetes}
\alias{Diabetes}
\title{Diabetes—A logistic data set, determining whether a woman tested positive for diabetes. 100 percent accurate results are possible using the logistic function in the Ensembles package.}
\format{
An object of class \code{data.frame} with 768 rows and 9 columns.
}
\source{
<https://www.kaggle.com/datasets/uciml/pima-indians-diabetes-database/data>
}
\usage{
Diabetes
}
\description{
"This dataset is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset
is to diagnostically predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset."

\describe{
This data set is from www.kaggle.com. The original notes on the website state:
Context
"This dataset is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset
is to diagnostically predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset.
Several constraints were placed on the selection of these instances from a larger database. In particular, all patients here are females
at least 21 years old of Pima Indian heritage."
Content
"The datasets consists of several medical predictor variables and one target variable, Outcome. Predictor variables includes the
number of pregnancies the patient has had, their BMI, insulin level, age, and so on.
Acknowledgements
Smith, J.W., Everhart, J.E., Dickson, W.C., Knowler, W.C., & Johannes, R.S. (1988). Using the ADAP learning algorithm to forecast the onset of diabetes mellitus.
In Proceedings of the Symposium on Computer Applications and Medical Care (pp. 261--265). IEEE Computer Society Press.

\item{Pregnancies}{Number of time pregnant}
\item{Glucose}{Plasma glucose concentration a 2 hours in an oral glucose tolerance test}
\item{BloodPressure}{Diastolic blood pressure (mm Hg)}
\item{SkinThickness}{Triceps skin fold thickness (mm)}
\item{Insulin}{2-Hour serum insulin (mu U/ml)}
\item{BMI}{Body mass index (weight in kg/(height in m)^2)}
\item{DiabetesPedigreeFunction}{Diabetes pedigree function}
\item{Age}{Age (years)}
\item{Outcome}{Class variable (0 or 1) 268 of 768 are 1, the others are 0}
}
}
\keyword{datasets}
