% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{German_Credit_Risk}
\alias{German_Credit_Risk}
\title{German_Credit_Risk-This dataset classifies people described by a set of attributes as good or bad credit risks.
#'}
\format{
An object of class \code{data.frame} with 999 rows and 21 columns.
}
\source{
https://archive.ics.uci.edu/dataset/144/statlog+german+credit+data
}
\usage{
German_Credit_Risk
}
\description{
This data set originally came from Professor Hofmann, and is available in several locations, including the UCI Machine Learning Repository
I cleaned the data set up, which included naming each of the columns, and removing white spaces from the names of the columns.
}
\keyword{datasets}
