% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{Logistic}
\alias{Logistic}
\title{logistic—function to perform logistic analysis and return the results to the user.}
\usage{
Logistic(
  data,
  colnum,
  numresamples,
  save_all_trained_models = c("Y", "N"),
  how_to_handle_strings = c("0", "1"),
  do_you_have_new_data = c("Y", "N"),
  remove_ensemble_correlations_greater_than,
  use_parallel = c("Y", "N"),
  train_amount,
  test_amount,
  validation_amount
)
}
\arguments{
\item{data}{data can be a CSV file or within an R package, such as MASS::Pima.te}

\item{colnum}{the column number with the logistic data}

\item{numresamples}{the number of resamples}

\item{save_all_trained_models}{"Y" or "N". Places all the trained models in the Environment}

\item{how_to_handle_strings}{0: No strings, 1: Factor values}

\item{do_you_have_new_data}{"Y" or "N". If "Y", then you will be asked for the new data}

\item{remove_ensemble_correlations_greater_than}{Enter a number to remove correlations in the ensembles}

\item{use_parallel}{"Y" or "N" for parallel processing}

\item{train_amount}{set the amount for the training data}

\item{test_amount}{set the amount for the testing data}

\item{validation_amount}{Set the amount for the validation data}
}
\value{
a real number
}
\description{
logistic—function to perform logistic analysis and return the results to the user.
}
\examples{
options(rgl.useNULL=TRUE) #needed if there are issues with X11 with this example
Logistic(data = SAHeart,
  colnum = 10,
  numresamples = 2,
  how_to_handle_strings = 1,
  do_you_have_new_data = "N",
  save_all_trained_models = "N",
  remove_ensemble_correlations_greater_than = 1.00,
  use_parallel = "N",
  train_amount = 0.60,
  test_amount = 0.20,
  validation_amount = 0.20)


options(rgl.useNULL=TRUE) #needed if there are issues with X11 with this example
Logistic(data = Lebron,
  colnum = 6,
  numresamples = 10,
  save_all_trained_models = "Y",
  how_to_handle_strings = 0,
  do_you_have_new_data = "N",
  remove_ensemble_correlations_greater_than = 1.00,
  use_parallel = "Y",
  train_amount = 0.50,
  test_amount = 0.25,
  validation_amount = 0.25)


options(rgl.useNULL=TRUE) #needed if there are issues with X11 with this example
Logistic(data = Diabetes,
   colnum = ncol(Diabetes),
   numresamples = 10,
   save_all_trained_models = "N",
   how_to_handle_strings = 1,
   do_you_have_new_data = "N",
   remove_ensemble_correlations_greater_than = 1.00,
   use_parallel = "Y",
   train_amount = 0.60,
   test_amount = 0.20,
   validation_amount = 0.20)

}
