% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAHeart}
\alias{SAHeart}
\title{SAHeart data}
\format{
SAHeart
\describe{
  \item{sbp}{Systolic blood pressure}
  \item{tobacco}{cumulative tobacco (kg)}
  \item{ldl}{low density lipoprotein cholesterol}
  \item{adiposity}{a numeric vector}
  \item{famhist}{family history of heart disease, a factor with levels Absent Present}
  \item{typea}{type-A behavior}
  \item{obesity}{a numeric vector}
  \item{alcohol}{current alcohol consumption}
  \item{age}{age at onset}
  \item{chd}{response, coronary heart disease}

}
}
\source{
Rousseauw, J., du Plessis, J., Benade, A., Jordaan, P., Kotze, J. and Ferreira, J. (1983). Coronary risk factor screening in three rural communities, South African Medical Journal 64: 430–436.
}
\usage{
SAHeart
}
\description{
This is the South African heart disease data originally published in Elements of Statistical Learning, see https://rdrr.io/cran/ElemStatLearn/man/SAheart.html
}
\keyword{datasets}
